+++ playwright.config.ts
@@ -21,7 +21,7 @@ export default defineConfig({
   /* Maximum time one test can run for. */
   timeout: 30 * 1000,
   expect: {
-    timeout: 10000
+    timeout: 20000
   },
   /* Run tests in files in parallel */
   fullyParallel: false,