+++ playwright/helpers/SearchFilters 3.ts
@@ -1,0 +1,10 @@
+export enum SearchFilters {
+  Epochs = "Epochs",
+  Blocks = "Blocks",
+  Transactions = "Transactions",
+  AdaHandle = "ADA Handle",
+  Tokens = "Tokens",
+  Addresses = "Addresses",
+  Pools = "Pools",
+  ScriptHash = "Script Hash"
+}