+++ src/components/DelegationPool/DelegationList/styles.ts
@@ -42,9 +42,10 @@ export const SearchContainer = styled("div")(({ theme }) => ({
 
 export const DelegationContainer = styled("div")(() => ({
   "& .table-wrapper tbody": {
-    zIndex: 10
+    zIndex: 1
   }
 }));
+
 export const StyledInput = styled("input")`
   border: none;
   width: 100%;